propgating devanagari characters to frontend.
